cambio de floats

+++ Engine/src/main/java/ucm/gdv/engine/Graphics.java
@@ -20,7 +20,7 @@ public interface Graphics {
 
     public void setColor(String color);
 
-    public void drawLine(int x1, int y1, int x2, int y2);
+    public void drawLine(float x1, float y1, float x2, float y2);
 
     public void fillRect(int x1, int y1, int x2, int y2);
